# Implement vectorized HdMergingSceneIndex::RemoveInputScenes()