--- a/pxr/imaging/hd/mergingSceneIndex.cpp
+++ b/pxr/imaging/hd/mergingSceneIndex.cpp
@@ -253,59 +253,76 @@ HdMergingSceneIndex::RemoveInputScenes(
 {
     TRACE_FUNCTION();
 
-    // Vectorization not implemented yet :(
-
-    for (const HdSceneIndexBaseRefPtr &sceneIndex : sceneIndices) {
-        RemoveInputScene(sceneIndex);
-    }
-}
-
-void
-HdMergingSceneIndex::RemoveInputScene(const HdSceneIndexBaseRefPtr &sceneIndex)
-{
-    TRACE_FUNCTION();
-
-    auto it = std::find_if(
-        _inputs.begin(), _inputs.end(),
-        [&sceneIndex](const _InputEntry &entry) {
-            return sceneIndex == entry.sceneIndex; });
-
-    if (it == _inputs.end()) {
+    if (sceneIndices.empty()) {
         return;
     }
 
-    std::vector<SdfPath> removalTestQueue = { it->sceneRoot };
+    // Remove our observer from the scene indices being removed.
+    HdSceneIndexObserverPtr observerPtr(&_observer);
+    for (const HdSceneIndexBaseRefPtr &sceneIndex : sceneIndices) {
+        sceneIndex->RemoveObserver(observerPtr);
+    }
 
-    sceneIndex->RemoveObserver(HdSceneIndexObserverPtr(&_observer));
-    _inputs.erase(it);
+    // Remove the scene indices from our list of inputs, and record a list of
+    // their scene roots for generating the added/removed notifications below.
+    std::vector<_InputEntry> removedInputs;
+    {
+        std::unordered_set<HdSceneIndexBaseRefPtr, TfHash> sceneIndicesSet;
+        sceneIndicesSet.insert(sceneIndices.begin(), sceneIndices.end());
+
+        auto it = std::stable_partition(
+            _inputs.begin(), _inputs.end(),
+            [&sceneIndicesSet](const _InputEntry &entry) {
+              return !sceneIndicesSet.count(entry.sceneIndex);
+            });
+
+        removedInputs.assign(it, _inputs.end());
+        _inputs.erase(it, _inputs.end());
+    }
+
     _RebuildInputsPathTable();
 
     if (!_IsObserved()) {
         return;
     }
 
-    // prims unique to this input get removed
+    // prims unique to these inputs get removed
     HdSceneIndexObserver::RemovedPrimEntries removedEntries;
 
-    // prims which this input contributed to are resynced via
+    // prims which these inputs contributed to are resynced via
     // PrimsAdded.
     HdSceneIndexObserver::AddedPrimEntries addedEntries;
 
-    // signal removal for anything not present once this scene is
-    // removed
-    while (!removalTestQueue.empty()) {
-        const SdfPath path = removalTestQueue.back();
-        removalTestQueue.pop_back();
+    // Set to prevent sending duplicate notifications.
+    std::unordered_set<SdfPath, SdfPath::Hash> visitedPaths;
 
-        const HdSceneIndexPrim prim = GetPrim(path);
-        if (!prim.dataSource
-                 && GetChildPrimPaths(path).empty()) {
-            removedEntries.emplace_back(path);
-        } else {
-            addedEntries.emplace_back(path, prim.primType);
-            for (const SdfPath &childPath :
-                     sceneIndex->GetChildPrimPaths(path)) {
-                removalTestQueue.push_back(childPath);
+    std::vector<SdfPath> removalTestQueue;
+    for (const _InputEntry &removedInput : removedInputs) {
+        // signal removal for anything not present once this scene is
+        // removed
+        removalTestQueue.push_back(removedInput.sceneRoot);
+
+        while (!removalTestQueue.empty()) {
+            const SdfPath path = removalTestQueue.back();
+            removalTestQueue.pop_back();
+
+            auto [_, notVisited] = visitedPaths.insert(path);
+
+            const HdSceneIndexPrim prim = GetPrim(path);
+            if (!prim.dataSource
+                     && GetChildPrimPaths(path).empty()) {
+                if (notVisited) {
+                    removedEntries.emplace_back(path);
+                }
+            } else {
+                if (notVisited) {
+                    addedEntries.emplace_back(path, prim.primType);
+                }
+
+                for (const SdfPath &childPath :
+                         removedInput.sceneIndex->GetChildPrimPaths(path)) {
+                    removalTestQueue.push_back(childPath);
+                }
             }
         }
     }
@@ -316,6 +333,12 @@ HdMergingSceneIndex::RemoveInputScene(const HdSceneIndexBaseRefPtr &sceneIndex)
     if (!addedEntries.empty()) {
         _SendPrimsAdded(addedEntries);
     }
+}
+
+void
+HdMergingSceneIndex::RemoveInputScene(const HdSceneIndexBaseRefPtr &sceneIndex)
+{
+    RemoveInputScenes({sceneIndex});
 }
 
 std::vector<HdSceneIndexBaseRefPtr>