### Description of Change(s)

This improves performance with some stage updates for instances, such as the attached example [removeinputscenes_example.zip](https://github.com/user-attachments/files/21621894/removeinputscenes_example.zip)

- With `USDIMAGINGGL_ENGINE_ENABLE_SCENE_INDEX=1`, run `usdview usdskel_instanced.usd`
- In the interpreter, run `usdviewApi.stage.GetSessionLayer().subLayerPaths.append("layer_materials.usda")`
- (example timings for the performance difference are listed below)

While implementing this I also noticed that the HdMergingSceneIndex test was never enabled, so this PR also re-enables it and adds a test for `RemoveInputScenes()` (the first commit adds a test for the original behaviour, with the second commit updating the expected result for the different notifications when removing a batch of scenes at once).

#### Timings

Snippets of timings before and after this change (note these timings also incorporated #3757 to improve the `_RebuildInputsPathTable()` performance)

Before:
```
    0.000 ms     0.000 ms       1 samples    |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.332 ms     0.332 ms   10000 samples    |   |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.000 ms     0.000 ms       1 samples    |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
161665.226 ms     3.050 ms       1 samples    |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
161662.176 ms   467.425 ms   10000 samples    |   |   |   |   |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScene
    0.301 ms     0.301 ms   10000 samples    |   |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
40771.477 ms     1.323 ms       1 samples    |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
40770.153 ms   199.747 ms   10000 samples    |   |   |   |   |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScene
    0.261 ms     0.261 ms   10000 samples    |   |   |   |   |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.000 ms     0.000 ms       1 samples    |   |   |   |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
  113.224 ms     0.016 ms       1 samples    |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScene
    0.009 ms     0.005 ms       1 samples    |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScene
```

After:
```
    0.000 ms     0.000 ms       1 samples    |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.548 ms     0.548 ms   10000 samples    |   |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.000 ms     0.000 ms       1 samples    |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
   67.336 ms    14.657 ms       1 samples    |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.437 ms     0.437 ms   10000 samples    |   |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
  124.699 ms    33.524 ms       1 samples    |   |   |   |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.370 ms     0.370 ms   10000 samples    |   |   |   |   |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.000 ms     0.000 ms       1 samples    |   |   |   |   |   |   pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
  109.827 ms     0.025 ms       1 samples    |   |   |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes
    0.015 ms     0.010 ms       1 samples    |   | pxrInternal_v0_25_8__pxrReserved__::HdMergingSceneIndex::RemoveInputScenes

```


### Link to proposal ([if applicable](https://openusd.org/release/contributing_to_usd.html#step-1-get-consensus-for-major-changes))
N/A

### Fixes Issue(s)
N/A

### Checklist

- [x] I have created this PR based on the dev branch

- [x] I have followed the [coding conventions](https://openusd.org/release/api/_page__coding__guidelines.html)

- [x] I have added unit tests that exercise this functionality (Reference: 
  [testing guidelines](https://openusd.org/release/api/_page__testing__guidelines.html))

- [x] I have verified that all unit tests pass with the proposed changes

- [x] I have submitted a signed Contributor License Agreement (Reference: 
  [Contributor License Agreement instructions](https://openusd.org/release/contributing_to_usd.html#contributor-license-agreement))